Draft WebClient life cycle management interface.
This will take care of instantiating/disposing of the browser window for clients that
- need the browser for a short time,
- then need the browser out of the way,
- may need to resume the window later.

An example of this is using the browser through a login gateway (Iovation blackbox, Cloudflare under attack mode, captcha...), then scraping the cookies and auth headers and continuing without the browser over e.g. https or wss.